+++ src/components/mainPanel/leftPanel/LeftPanel.svelte
@@ -1,17 +1,13 @@
 <script lang="ts">
-  import { onMount } from 'svelte';
   import { rootNodes } from '../../../store';
-  let timeOver = false;
-  import DummyTree2 from './dummyTree2.svelte';
-  setTimeout(() => (timeOver = !timeOver), 10000);
-  rootNodes.subscribe(() => console.log($rootNodes, Date.now()));
+  import Tree from './Tree.svelte';
 </script>
 
 <!-- check to see if rootNodes is undefined or not -->
 
 {#if $rootNodes.length}
   <div class="container">
-    <DummyTree2 />
+    <Tree />
   </div>
 {:else}
   <div class="container">not yet ready</div>
